Moved unnecessary lib functions to their respective files
