+++ bot/commands/todo/lib.go
@@ -221,141 +221,6 @@ func (s Todo) changeItemsStatus(userId string, itemIds []string, from, to string
 	return nil
 }
 
-// Adds an active todo item
-func (s Todo) addItem(author, title, description string) {
-	db, _ := sql.Open("postgres", s.PsqlConn)
-	defer db.Close()
-
-	taskId := 0
-	rows, _ := db.Query(`SELECT MAX(id) FROM todo.task`)
-	if rows.Next() {
-		rows.Scan(&taskId)
-		rows.Close()
-		taskId++
-	}
-
-	// Insert task into task table
-	db.Exec(
-		`INSERT INTO todo.task (id, creator, title, description) VALUES ($1, $2, $3, $4)`,
-		taskId,
-		author,
-		title,
-		description,
-	)
-
-	// Insert task into active
-	db.Exec(
-		`INSERT INTO todo.active (discord_user, task) VALUES ($1, $2)`,
-		author,
-		taskId,
-	)
-}
-
-// Deletes todo items from the user
-func (s Todo) deleteItems(userId string, items []string) error {
-	userItems, err := s.getAllTodos(userId)
-	if err != nil {
-		return err
-	}
-
-	// For checking for invalid IDs
-	itemsCopy := make([]string, len(items))
-	copy(itemsCopy, items)
-
-	for _, item := range userItems {
-		for i := range itemsCopy {
-			if itemsCopy[i] == fmt.Sprint(item.ID) {
-				itemsCopy = append(itemsCopy[:i], itemsCopy[i+1:]...)
-				break
-			}
-		}
-	}
-
-	// Check for wrong ID supplied
-	if len(itemsCopy) != 0 {
-		return fmt.Errorf("user %s has no task with id %s", userId, strings.Join(itemsCopy, ", "))
-	}
-
-	db, err := sql.Open("postgres", s.PsqlConn)
-	if err != nil {
-		return err
-	}
-	defer db.Close()
-
-	// Delete items
-	for _, table := range []string{"active", "completed", "archived"} {
-		_, err = db.Exec(fmt.Sprintf(`DELETE FROM todo.%s WHERE discord_user=$1 AND task=any($2)`, table),
-			userId,
-			pq.Array(items),
-		)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-// Archives active/completed items from the user
-func (s Todo) archiveItems(userId string, items []string) error {
-	active, err := s.getActiveTodos(userId)
-	if err != nil {
-		return err
-	}
-	completed, err := s.getDoneTodos(userId)
-	if err != nil {
-		return err
-	}
-
-	userItems := append(active, completed...)
-
-	// For checking for invalid IDs
-	itemsCopy := make([]string, len(items))
-	copy(itemsCopy, items)
-
-	for _, item := range userItems {
-		for i := range itemsCopy {
-			if itemsCopy[i] == fmt.Sprint(item.ID) {
-				itemsCopy = append(itemsCopy[:i], itemsCopy[i+1:]...)
-				break
-			}
-		}
-	}
-
-	// Check for wrong ID supplied
-	if len(itemsCopy) != 0 {
-		return fmt.Errorf("user %s has no active or completed task with id %s", userId, strings.Join(itemsCopy, ", "))
-	}
-
-	db, err := sql.Open("postgres", s.PsqlConn)
-	if err != nil {
-		return err
-	}
-	defer db.Close()
-
-	// Delete items
-	for _, table := range []string{"active", "completed"} {
-		_, err = db.Exec(fmt.Sprintf(`DELETE FROM todo.%s WHERE discord_user=$1 AND task=any($2)`, table),
-			userId,
-			pq.Array(items),
-		)
-		if err != nil {
-			return err
-		}
-	}
-
-	// Put all items into archived
-	_, err = db.Exec(`INSERT INTO todo.archived (discord_user, task) VALUES ($1, UNNEST($2::INTEGER[]))`,
-		userId,
-		pq.Array(items),
-	)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
 // Sends a message with the option for the user to select multiple items at once
 // If the user presses the green button, submit gets called
 // If the user presses the red button, cancel gets called
